requirements.txt
Django==4.1.5
django-cleanup
django-extensions
Pillow
tqdm
